+++ source/databricks/settlement_report/settlement_report_job/domain/report_generator.py
@@ -5,7 +5,7 @@ from settlement_report_job.domain.time_series_factory import create_time_series
 
 from settlement_report_job.utils import create_zip_file, get_dbutils
 from settlement_report_job.constants import get_output_volume_name
-from settlement_report_job.logger import Logger
+from settlement_report_job.infrastructure.logger import Logger
 
 log = Logger(__name__)
