+++ source/databricks/settlement_report/settlement_report_job/infrastructure/table_column_names.py
@@ -19,3 +19,19 @@ class EnergyResultsCsvColumnNames:
     metering_point_type = "TYPEOFMP"
     settlement_method = "SETTLEMENTMETHOD"
     quantity = "ENERGYQUANTITY"
+
+
+class TimeSeriesPointCsvColumnNames:
+    metering_point_id = "METERINGPOINTID"
+    metering_point_type = "TYPEOFMP"
+    start_of_day = "STARTDATETIME"
+    energy_prefix = "ENERGYQUANTITY"
+
+
+class EphemeralColumns:
+    """Columns that are added to the DataFrame for processing but not part of the input
+    or output schema."""
+
+    uid = "uid"
+    start_of_day = "start_of_day"
+    quantities = "quantities"
