--- source/databricks/settlement_report/tests/domain/energy_v1.csv
@@ -1,6 +1,0 @@
-calculation_id;calculation_type;calculation_period_start;calculation_period_end;calculation_version;result_id;grid_area_code;metering_point_type;settlement_method;resolution;time;quantity;quantity_unit;quantity_qualities
-0b15a420-9fc8-409a-a169-fbd49479d718;wholesale_fixing;2023-01-01 00:00:00;2023-02-01 00:00:00;1;11111111-9fc8-409a-a169-fbd49479d718;804;production;;PT15M;2023-02-02 02:00;1.125;kWh;['measured']
-0b15a420-9fc8-409a-a169-fbd49479d718;wholesale_fixing;2023-01-01 00:00:00;2023-02-01 00:00:00;1;22222222-9fc8-409a-a169-fbd49479d718;804;consumption;non_profiled;PT15M;2023-02-02 02:00;1.125;kWh;['measured']
-0b15a420-9fc8-409a-a169-fbd49479d718;wholesale_fixing;2023-01-01 00:00:00;2023-02-01 00:00:00;1;33333333-9fc8-409a-a169-fbd49479d718;804;consumption;flex;PT15M;2023-02-02 02:00;1.125;kWh;['measured']
-0b15a420-9fc8-409a-a169-fbd49479d718;wholesale_fixing;2023-01-01 00:00:00;2023-02-01 00:00:00;1;44444444-9fc8-409a-a169-fbd49479d718;804;consumption;;PT15M;2023-02-02 02:00;1.125;kWh;['measured']
-0b15a420-9fc8-409a-a169-fbd49479d718;wholesale_fixing;2023-01-01 00:00:00;2023-02-01 00:00:00;1;55555555-9fc8-409a-a169-fbd49479d718;804;exchange;;PT15M;2023-02-02 02:00;1.125;kWh;['measured']
